Add files via upload
a.	Create an php file (called it order) to represents an order form that looks like this.
b.	Create a php document that called process_order.php. Link this file to the action attribute in previous file (order) so that when the user clicks submit button, the code in process_order.php gets executed. 
c.	Now implement process_order,php file to read and print out quantity values of Tires, Oil and Spark Plugs you supposed to enter in the order page.